requirements/base.txt
Django==1.11.1
PyJWT==1.4.2
djangorestframework==3.6.3
djangorestframework-jwt==1.8.0
wheel==0.24.0